Add Auth to USER to manage CORS request
JWT token is managing the authenticartion process
through the gem Knock

The contest controller need auth to be accessed so please
login first and copy the token auth returned by JWT in to
the Bearer Header for following CORS requests.

+++ db/migrate/20191011133813_create_songs.rb
@@ -5,7 +5,7 @@ class CreateSongs < ActiveRecord::Migration[5.2]
       t.string :artist
       t.string :album
       t.string :spotify_id
-      t.float :length_ms
+      t.integer :length_ms
 
       t.timestamps
     end
